+++ src/update/models/firmwareManifest.ts
@@ -34,10 +34,6 @@ export class FirmwareManifest {
      */
     public readonly url: string;
     /**
-     * Checksum generated for the datafile
-     */
-    public readonly datafileChecksum?: string;
-    /**
      * Size of the datafile (in bytes)
      */
     public readonly datafileSize?: number;
@@ -46,11 +42,11 @@ export class FirmwareManifest {
      */
     public readonly deviceClass?: string;
     /**
-     * The time the object was created
+     * The timestamp when the object was created
      */
     public readonly createdAt: Date;
     /**
-     * The time the object was updated
+     * The timestamp the object was updated
      */
     public readonly updatedAt: Date;
     /**