+++ src/update/updateApi.ts
@@ -771,7 +771,7 @@ export class UpdateApi {
     public updateCampaign(campaign: UpdateCampaignObject, callback: CallbackFn<Campaign>): void;
     public updateCampaign(campaign: UpdateCampaignObject, callback?: CallbackFn<Campaign>): Promise<Campaign> {
         return apiWrapper(resultsFn => {
-            this._endpoints.update.updateCampaignPartialUpdate(campaign.id, CampaignAdapter.updateMap(campaign), resultsFn);
+            this._endpoints.update.updateCampaignUpdate(campaign.id, CampaignAdapter.updateMap(campaign), resultsFn);
         }, (data, done) => {
             const response = CampaignAdapter.map(data, this);
             done(null, response);
@@ -951,7 +951,7 @@ export class UpdateApi {
 
         return apiWrapper(resultsFn => {
             const { limit, order, after, include } = options;
-            this._endpoints.update.v3UpdateCampaignsCampaignIdCampaignDeviceMetadataGet(campaignId, limit, order, after, encodeInclude(include), resultsFn);
+            this._endpoints.update.updateCampaignMetadataList(campaignId, limit, order, after, encodeInclude(include), resultsFn);
         }, (data, done) => {
             let list: Array<CampaignDeviceState>;
             if (data.data && data.data.length) {